requirements.txt
google-cloud-texttospeech==2.14.1
openai==0.27.0
SpeechRecognition==3.9.0